Fix bash globbing

+++ install.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 # Copy stuff to the right location
 
-cp HourlyWallpaper $HOME/.config/
+if [[ ! -f "$HOME/.config/HourlyWallpaper" ]]; then 
+	cp HourlyWallpaper $HOME/.config/
+fi
 cp wallchanger $HOME/.local/bin/
 
 sudo cp wallchanger.cron /etc/cron.d/wallchanger
+
